pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "auto-improvement"
version = "0.1.0"
description = "Auto-improvement system for Claude/Cursor/Codex that learns from real PRs"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Auto Improvement Contributors"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "gitpython>=3.1.40",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-vcr>=1.0.2",
    "vcrpy>=6.0.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.8.0",
    "mypy>=1.5.0",
    "types-pyyaml>=6.0.0",
    "types-requests>=2.31.0",
    "types-beautifulsoup4>=4.12.0",
]

[project.scripts]
auto-improve = "auto_improvement.cli:app"

# Ruff configuration (replaces black + flake8 + isort)
[tool.ruff]
line-length = 100
target-version = "py312"

# Enable additional rules
[tool.ruff.lint]
select = [
    "B",   # flake8-bugbear
    "D",   # pydocstyle
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "PL",  # pylint
    "UP",  # pyupgrade
    "RUF1", # ruff-specific, noqa codes
    "W",   # pycodestyle warnings
]
ignore = [
    "E501",    # line too long (handled by formatter)
    "D104",    # missing docstring in public package
    "D107",    # missing docstring in __init__
    "D102",    # missing docstring in public method
    "D401",    # first line should be in imperative mood
    "D400",    # first line should end with a period
    "D415",    # first line should end with punctuation
    "D417",    # missing argument descriptions in docstring
    "PLR2004", # magic value used in comparison
    "SIM115",  # use context manager for opening files
    "D212",    # multi-line-summary-first-line
    "D203",    # one-blank-line-before-class, incompatible with D211
    "PLC0415", # import not at top of file (intentional for lazy loading to avoid circular deps)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# Mypy configuration (strict mode)
[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_unreachable = true
no_implicit_reexport = true
strict_equality = true

# Allow some flexibility for external libraries
[[tool.mypy.overrides]]
module = [
    "git.*",
    "anthropic.*",
    "typer.*",
    "rich.*",
]
ignore_missing_imports = true

# Typer decorators are untyped, allow in CLI module
[[tool.mypy.overrides]]
module = "auto_improvement.cli"
disallow_untyped_decorators = false

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20250915",
]
